Add (private) shallow_clone method to request
To resolve #1116 and #1245, we need to mess about with $env entries used
by Plack::Request, which shouldn't be done in the App class

Refactor the code from Dancer2::Core::App->make_forward_to, moving all
the request related code (which was the bulk of it) into the request
class.

For #1245, extra params are merged into `params`, `query_params` and
`query_parameters`.

For #1116, ensure we clone the previous HVM in
$env->{plack.request.body} into the new request.